# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vmedis Proxy API
  description: |
    Proxy API for Vmedis, providing pharmacy-related APIs such as sales, drugs,
    procurements, stock opnames, shifts, rejected drugs, and Vmedis token management.

    ## Authentication
    Authentication is done by sending the user's email address in the `X-Email` header.
    Requests without the header are treated as the `guest` user. Some endpoints
    require the `admin` or `staff` role and respond with `403` otherwise.

    ## API versions
    - `/api/v1`: raw data endpoints.
    - `/api/v2`: endpoints where the response is generated on the server side and
      authentication might affect the response.

    ## Time range query parameters
    Endpoints that accept a time range use the following query parameters
    (all dates use the `YYYY-MM-DD` format):
    - `date`: a single day.
    - `from` + `until` (alias `to`): an inclusive day range. When only `from` is
      given, the range ends at the end of today.
    - When no parameter is given, the range defaults to today.
  version: 1.0.0
servers:
  - url: /
    description: This server.
...
security:
  - {}
  - EmailAuth: []
paths:
...
  /api/v2/shifts:
    get:
      operationId: getShifts
      tags: [Shifts]
      summary: Get shifts
      description: |
        Returns the cashier shifts in the given time range (defaults to today)
        as a display-ready table. Requires the `admin` or `staff` role.
      security:
        - EmailAuth: []
      parameters:
        - $ref: '#/components/parameters/DateQuery'
        - $ref: '#/components/parameters/FromQuery'
        - $ref: '#/components/parameters/UntilQuery'
        - $ref: '#/components/parameters/ToQuery'
      responses:
        '200':
          description: Shifts in the given time range as a table.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v2/shifts/dump/status:
    get:
      operationId: getShiftDumpStatus
      tags: [Shifts]
      summary: Get shift dump status
      description: |
        Returns whether a shift dump is currently in progress.
        `DUMPING` means a dump is running (the distributed lock is held);
        `IDLE` means no dump is running. Requires the `admin` or `staff` role.
      security:
        - EmailAuth: []
      responses:
        '200':
          description: Shift dump status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftDumpStatusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/shifts/{vmedis_id}:
    get:
      operationId: getShiftByVmedisID
      tags: [Shifts]
      summary: Get shift by Vmedis ID
      description: |
        Returns the shift with the given Vmedis ID as a display-ready table of
        label/value rows. Requires the `admin` or `staff` role.
      security:
        - EmailAuth: []
      parameters:
        - $ref: '#/components/parameters/ShiftVmedisID'
      responses:
        '200':
          description: The shift as a table.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v2/shifts/{vmedis_id}/show:
    get:
      operationId: showShift
      tags: [Shifts]
      summary: Show shift as HTML
      description: |
        Renders the shift with the given Vmedis ID as an HTML page.
        Requires the `admin` or `staff` role.
      security:
        - EmailAuth: []
      parameters:
        - $ref: '#/components/parameters/ShiftVmedisID'
      responses:
        '200':
          description: The shift rendered as an HTML page.
          content:
            text/html:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    EmailAuth:
      type: apiKey
      in: header
      name: X-Email
      description: |
        The email address of the user. Requests without this header are treated
        as the `guest` user.
  parameters:
    DateQuery:
      name: date
      in: query
      description: A single day, in `YYYY-MM-DD` format. Ignored when `from` is given.
      schema:
        type: string
        format: date
    FromQuery:
      name: from
      in: query
      description: Start of the day range, in `YYYY-MM-DD` format. Takes precedence over `date`.
      schema:
        type: string
        format: date
    UntilQuery:
      name: until
      in: query
      description: End (inclusive) of the day range, in `YYYY-MM-DD` format. Only used together with `from`. Defaults to today.
      schema:
        type: string
        format: date
    ToQuery:
      name: to
      in: query
      description: Alias of `until`. Ignored when `until` is given.
      schema:
        type: string
        format: date
    ShiftVmedisID:
      name: vmedis_id
      in: path
      required: true
      description: The Vmedis ID of the shift.
      schema:
        type: integer
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The user's role is not allowed to access this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A human-readable error message.
      required: [error]
...
    ShiftDumpStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [IDLE, DUMPING]
          description: |
            Whether a shift dump is currently in progress.
            `DUMPING` means a dump is running; `IDLE` means none is running.

    # ----- Generic display components (cui) -----
    Table:
      type: object
      description: A generic display-ready table.
      properties:
        header:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            $ref: '#/components/schemas/TableRow'
    TableRow:
      type: object
      properties:
        id:
          type: string
        columns:
          type: array
          items:
            type: string
...